refactor(python): keep sentinel_for_numpy_dtype out of public lib/constants/
Internal's own split puts only the raw sentinel VALUES in the public
shared/constants/sentinel.py; the dtype-dispatch function
sentinel_for_numpy_dtype stays in shared/utils/numpy_utils/sentinel.py,
importing the public constants. The prior commit (5050a05d) over-broadened
the promotion by moving the function itself into lib/constants/. Move it
back to _internal/utils/numpy_utils/sentinel.py, matching internal's actual
public/private boundary: only the framework-agnostic values are public;
the numpy-coupled dispatch logic is not.
